move camera stream definitions into cereal
