+++ openpilot/cereal/visionstream.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "msgq/visionipc/visionbuf.h"
+
+enum VisionStreamValues : VisionStreamType {
+  VISION_STREAM_ROAD = 0,
+  VISION_STREAM_DRIVER = 1,
+  VISION_STREAM_WIDE_ROAD = 2,
+  VISION_STREAM_MAP = 3,
+};
