+++ openpilot/tools/replay/camera.h
@@ -5,7 +5,7 @@
 #include <tuple>
 #include <utility>
 
-#include "common/visionstream.h"
+#include "openpilot/cereal/visionstream.h"
 #include "msgq/visionipc/visionipc_server.h"
 #include "common/queue.h"
 #include "tools/replay/framereader.h"
